deliver the get data and clear data functions from hook
